Decouple instructions into a separate feature flag.
