+++ README.md
@@ -8,7 +8,8 @@ Support for x86_64 specific instructions (e.g. TLB flush), registers (e.g. contr
 ## Crate Feature Flags
 
 * `nightly`: This is the default.
-* `stable`: Use this to build with non-nightly rust. Needs `default-features = false`.
+* `instructions`: Enabled by default, turns on x86\_64 specific instructions, and dependent features. Only available for x86\_64 targets.
+* `external_asm`: Use this to build with non-nightly rust. Needs `default-features = false, features = ["instructions"]`. Is unsupported on Windows.
 
 ## Building with stable rust
