+++ src/structures/paging/mapper/mod.rs
@@ -1,8 +1,9 @@
 //! Abstractions for reading and modifying the mapping of pages.
 
 pub use self::mapped_page_table::{MappedPageTable, PhysToVirt};
-#[cfg(target_arch = "x86_64")]
-pub use self::{offset_page_table::OffsetPageTable, recursive_page_table::RecursivePageTable};
+pub use self::offset_page_table::OffsetPageTable;
+#[cfg(feature = "instructions")]
+pub use self::recursive_page_table::RecursivePageTable;
 
 use crate::structures::paging::{
     frame_alloc::FrameAllocator, page_table::PageTableFlags, Page, PageSize, PhysFrame, Size1GiB,
@@ -12,6 +13,7 @@ use crate::{PhysAddr, VirtAddr};
 
 mod mapped_page_table;
 mod offset_page_table;
+#[cfg(feature = "instructions")]
 mod recursive_page_table;
 
 /// This trait defines page table operations that work for all page sizes of the x86_64
@@ -355,7 +357,7 @@ impl<S: PageSize> MapperFlush<S> {
     }
 
     /// Flush the page from the TLB to ensure that the newest mapping is used.
-    #[cfg(target_arch = "x86_64")]
+    #[cfg(feature = "instructions")]
     #[inline]
     pub fn flush(self) {
         crate::instructions::tlb::flush(self.0.start_address());
@@ -383,7 +385,7 @@ impl MapperFlushAll {
     }
 
     /// Flush all pages from the TLB to ensure that the newest mapping is used.
-    #[cfg(target_arch = "x86_64")]
+    #[cfg(feature = "instructions")]
     #[inline]
     pub fn flush_all(self) {
         crate::instructions::tlb::flush_all()
