+++ src/structures/paging/mod.rs
@@ -8,10 +8,12 @@ pub use self::frame_alloc::UnusedPhysFrame;
 pub use self::frame_alloc::{FrameAllocator, FrameDeallocator};
 #[doc(no_inline)]
 pub use self::mapper::MappedPageTable;
-pub use self::mapper::{Mapper, MapperAllSizes};
-#[cfg(target_arch = "x86_64")]
 #[doc(no_inline)]
-pub use self::mapper::{OffsetPageTable, RecursivePageTable};
+pub use self::mapper::OffsetPageTable;
+#[cfg(feature = "instructions")]
+#[doc(no_inline)]
+pub use self::mapper::RecursivePageTable;
+pub use self::mapper::{Mapper, MapperAllSizes};
 pub use self::page::{Page, PageSize, Size1GiB, Size2MiB, Size4KiB};
 pub use self::page_table::{PageOffset, PageTable, PageTableFlags, PageTableIndex};
